[pc-61] Relocate some files in a more cohesive way
Some reviewers pointed out the fact that some files were misplaced in the project tree. Some files have been relocated and/or renamed in order to keep a cohesive project tree.
