+++ api/lib/domain/services/session-xml-service.js
@@ -1,9 +1,9 @@
-const xmlService = require('./xml-service');
+const odsContentXmlService = require('./../../infrastructure/services/ods-content-xml-service');
 // Placeholder in the template ODS file that helps us find the template candidate row in the file.
 const CANDIDATE_ROW_MARKER_PLACEHOLDER = 'COUNT';
 
 function getUpdatedXmlWithSessionData({ stringifiedXml, sessionTemplateValues, sessionData }) {
-  return xmlService.updateXmlSparseValues({
+  return odsContentXmlService.updateXmlSparseValues({
     stringifiedXml,
     templateValues: sessionTemplateValues,
     dataToInject: sessionData,
@@ -11,7 +11,7 @@ function getUpdatedXmlWithSessionData({ stringifiedXml, sessionTemplateValues, s
 }
 
 function getUpdatedXmlWithCertificationCandidatesData({ stringifiedXml, candidateTemplateValues, candidatesData }) {
-  return xmlService.updateXmlRows({
+  return odsContentXmlService.updateXmlRows({
     stringifiedXml,
     rowMarkerPlaceholder: CANDIDATE_ROW_MARKER_PLACEHOLDER,
     rowTemplateValues: candidateTemplateValues,
